quant-stream/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiosignal==1.4.0
alembic==1.17.2
-e alphacopilot
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
args==0.1.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
authlib==1.6.5
beartype==0.15.0
beautifulsoup4==4.14.2
billiard==4.2.3
bleach==6.3.0
blinker==1.9.0
bokeh==3.8.1
boto3==1.40.76
botocore==1.40.76
cachetools==6.2.2
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
clint==0.5.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
coverage==7.12.0
cryptography==46.0.3
curl-cffi==0.13.0
cycler==0.12.1
cyclopts==4.2.4
databricks-sdk==0.73.0
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
deltalake==0.17.4
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring-parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.121.2
fastmcp==2.12.4
filelock==3.20.0
fire==0.7.1
flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
frozenlist==1.8.0
fs==2.4.16
fsspec==2025.10.0
genbadge==1.1.2
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
gitpython==3.1.45
google==3.0.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-bigquery==3.29.0
google-cloud-core==2.5.0
google-cloud-pubsub==2.33.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
groq==0.34.1
growwapi==1.3.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h3==4.3.1
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huey==2.5.4
idna==3.11
importlib-metadata==8.7.0
iniconfig==2.3.0
ipython==9.7.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter-bokeh==4.0.5
jupyterlab-widgets==3.0.16
kiwisolver==1.4.9
kombu==5.5.4
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-groq==1.0.1
langchain-mcp-adapters==0.1.13
langchain-openai==1.0.3
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.43
lazy-object-proxy==1.12.0
lightgbm==4.6.0
linkify-it-py==2.0.3
lxml==6.0.2
mako==1.3.10
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mcp==1.21.2
mdit-py-plugins==0.5.0
mdurl==0.1.2
mlflow==3.6.0
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy-extensions==1.1.0
narwhals==2.12.0
nats-py==2.12.0
networkx==3.5
nkeys==0.2.1
numpy==1.26.4
openai==2.8.1
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
panel==1.8.3
param==2.3.0
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pathway==0.27.0
pexpect==4.9.0
pillow==11.1.0
pluggy==1.6.0
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==18.1.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pynacl==1.6.1
pyotp==2.9.0
pyparsing==3.2.5
pyperclip==1.11.0
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-sat==1.8.dev24
pytz==2025.2
pyviz-comms==3.0.6
pyyaml==6.0.3
-e .
redis==5.2.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-futures==1.0.2
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.29.0
rsa==4.9.1
ruff==0.14.5
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlalchemy==2.0.44
sqlparse==0.5.3
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
werkzeug==3.1.1
widgetsnbextension==4.0.15
xgboost==3.1.1
xxhash==3.6.0
xyzservices==2025.10.0
yahooquery==2.4.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
torchvision==0.24.1